requirements.txt
feedparser
json
datetime
dateparser
nltk
os
tweepy
Pillow
configparser